Safari-611.1.10/LayoutTests/http/tests/inspector/network/resource-response-source-network-expected.txt

Test for `Resource.ResponseSource.Network`.


== Running test suite: Resource.ResponseSource.Network
-- Running test case: Resource.ResponseSource.Network
PASS: Resource should be created.
PASS: Resource should receive a Response.
PASS: statusCode should be 200
PASS: responseSource should be Symbol(network)
